Folders were renamed
