+++ tms-ts/src/App.tsx
@@ -2,10 +2,10 @@ import React from "react";
 import './App.css';
 import {Route, Routes, BrowserRouter} from "react-router-dom";
 import SuitesComponent from "./components/testcases/suites.component";
-import ProjectPage from "./pages/ProjectPage";
+import ProjectPage from "./components/projects/ProjectPage";
 import Header from "./components/Header";
 import Login from "./components/Login"
-import ProjectSelectionPage from "./pages/ProjectSelectionPage";
+import ProjectSelectionPage from "./components/projects/ProjectSelectionPage";
 
 
 function App() {
